fix(telegram): simplify thinking defaults
