+++ extensions/telegram/src/bot-native-commands.ts
@@ -2,11 +2,10 @@ import { randomUUID } from "node:crypto";
 import path from "node:path";
 import type { Bot, Context } from "grammy";
 import {
-  buildConfiguredModelCatalog,
   loadModelCatalog,
   resolveAgentConfig,
   resolveDefaultModelForAgent,
-  resolveThinkingDefault,
+  resolveThinkingDefaultWithRuntimeCatalog,
 } from "openclaw/plugin-sdk/agent-runtime";
 import { resolveChannelStreamingBlockEnabled } from "openclaw/plugin-sdk/channel-streaming";
 import { resolveNativeCommandSessionTargets } from "openclaw/plugin-sdk/command-auth-native";
@@ -262,34 +261,11 @@ async function resolveTelegramDefaultThinkingLevel(params: {
   provider: string;
   model: string;
 }): Promise<string> {
-  const configuredCatalog = buildConfiguredModelCatalog({ cfg: params.cfg });
-  const configuredSelectedEntry = configuredCatalog.find(
-    (entry) => entry.provider === params.provider && entry.id === params.model,
-  );
-  const shouldHydrateRuntimeCatalog =
-    configuredCatalog.length === 0 ||
-    !configuredSelectedEntry ||
-    configuredSelectedEntry.reasoning === undefined;
-  let runtimeCatalog: Awaited<ReturnType<typeof loadModelCatalog>> | undefined;
-  if (shouldHydrateRuntimeCatalog) {
-    try {
-      runtimeCatalog = await loadModelCatalog({ config: params.cfg });
-    } catch {
-      runtimeCatalog = undefined;
-    }
-  }
-  const runtimeSelectedEntry = runtimeCatalog?.find(
-    (entry) => entry.provider === params.provider && entry.id === params.model,
-  );
-  const catalog =
-    runtimeSelectedEntry || configuredCatalog.length === 0
-      ? (runtimeCatalog ?? configuredCatalog)
-      : configuredCatalog;
-  return resolveThinkingDefault({
+  return resolveThinkingDefaultWithRuntimeCatalog({
     cfg: params.cfg,
     provider: params.provider,
     model: params.model,
-    catalog,
+    loadModelCatalog: () => loadModelCatalog({ config: params.cfg }),
   });
 }
