+++ src/agents/model-thinking-default.ts
@@ -7,6 +7,7 @@ import {
 import type { ModelCatalogEntry } from "./model-catalog.types.js";
 import { legacyModelKey, modelKey, normalizeProviderId } from "./model-selection-normalize.js";
 import { normalizeModelSelection } from "./model-selection-resolve.js";
+import { buildConfiguredModelCatalog } from "./model-selection-shared.js";
 
 type ThinkLevel = "off" | "minimal" | "low" | "medium" | "high" | "xhigh" | "adaptive" | "max";
 
@@ -75,5 +76,35 @@ export function resolveThinkingDefault(params: {
     provider: params.provider,
     model: params.model,
     catalog: params.catalog,
+  });
+}
+
+export async function resolveThinkingDefaultWithRuntimeCatalog(params: {
+  cfg: OpenClawConfig;
+  provider: string;
+  model: string;
+  loadModelCatalog: () => Promise<ModelCatalogEntry[]>;
+}): Promise<ThinkLevel> {
+  const configuredCatalog = buildConfiguredModelCatalog({ cfg: params.cfg });
+  const configuredSelectedEntry = configuredCatalog.find(
+    (entry) => entry.provider === params.provider && entry.id === params.model,
+  );
+  const needsRuntimeCatalog =
+    configuredCatalog.length === 0 ||
+    !configuredSelectedEntry ||
+    configuredSelectedEntry.reasoning === undefined;
+  const runtimeCatalog = needsRuntimeCatalog ? await params.loadModelCatalog() : undefined;
+  const runtimeSelectedEntry = runtimeCatalog?.find(
+    (entry) => entry.provider === params.provider && entry.id === params.model,
+  );
+  const catalog =
+    runtimeSelectedEntry || configuredCatalog.length === 0
+      ? (runtimeCatalog ?? configuredCatalog)
+      : configuredCatalog;
+  return resolveThinkingDefault({
+    cfg: params.cfg,
+    provider: params.provider,
+    model: params.model,
+    catalog,
   });
 }
